moved blockList and blackListed URLs to ruleMatcher
