+++ browser-extension/mv3/src/service-worker/services/clientHandler.ts
@@ -1,8 +1,9 @@
-import { generateUrlPattern, getBlockedDomains, isExtensionEnabled, onBlockListChange } from "../../utils";
+import { generateUrlPattern, isExtensionEnabled } from "../../utils";
 import { WEB_URL, OTHER_WEB_URLS } from "../../../../config/dist/config.build.json";
 import { Variable, onVariableChange } from "../variable";
 import { RuleType } from "common/types";
 import rulesStorageService from "../../rulesStorageService";
+import { getBlockedDomains, onBlockListChange } from "../../common/ruleMatcher";
 
 const excludeMatchesPatterns = [WEB_URL, ...OTHER_WEB_URLS].map(generateUrlPattern).filter((pattern) => !!pattern);
