+++ browser-extension/mv3/src/service-worker/services/rulesManager.ts
@@ -1,12 +1,13 @@
 import config from "common/config";
 import { getEnabledRules, onRuleOrGroupChange } from "common/rulesStore";
 import { onVariableChange, Variable } from "../variable";
-import { debounce, getBlockedDomains, isExtensionEnabled, onBlockListChange } from "../../utils";
+import { debounce, isExtensionEnabled } from "../../utils";
 import { TAB_SERVICE_DATA, tabService } from "./tabService";
 import { SessionRuleType } from "./requestProcessor/types";
 import { sendMessageToApp } from "./messageHandler";
 import { EXTENSION_MESSAGES } from "common/constants";
 import { UpdateDynamicRuleOptions } from "common/types";
+import { getBlockedDomains, onBlockListChange } from "../../common/ruleMatcher";
 
 const ALL_RESOURCE_TYPES = Object.values(chrome.declarativeNetRequest.ResourceType);
