+++ browser-extension/mv3/src/utils.ts
@@ -1,9 +1,5 @@
-import { SourceKey, SourceOperator, UrlSource } from "common/types";
 import config from "common/config";
-import { matchSourceUrl } from "./common/ruleMatcher";
 import { Variable, getVariable } from "./service-worker/variable";
-import { ChangeType, getRecord, onRecordChange } from "common/storage";
-import { STORAGE_KEYS } from "common/constants";
 
 export const formatDate = (dateInMillis: number, format: string): string => {
   if (dateInMillis && format === "yyyy-mm-dd") {
@@ -27,23 +23,6 @@ export const getAllSupportedWebURLs = () => {
 
 export const isAppURL = (url: string) => {
   return !!url && getAllSupportedWebURLs().some((webURL) => url.includes(webURL));
-};
-
-export const isBlacklistedURL = (url: string): boolean => {
-  const blacklistedSources: UrlSource[] = [
-    ...getAllSupportedWebURLs().map((webUrl) => ({
-      key: SourceKey.URL,
-      operator: SourceOperator.CONTAINS,
-      value: webUrl,
-    })),
-    {
-      key: SourceKey.URL,
-      operator: SourceOperator.CONTAINS,
-      value: "__rq", // you can use __rq in the url to blacklist it
-    },
-  ];
-
-  return blacklistedSources.some((source) => matchSourceUrl(source, url));
 };
 
 export const generateUrlPattern = (urlString: string) => {
@@ -77,48 +56,4 @@ export const debounce = (func: Function, wait: number) => {
     clearTimeout(timeout);
     timeout = setTimeout(() => func(...args), wait);
   };
-};
-
-let cachedBlockedDomains: string[] | null = null;
-
-export const cacheBlockedDomains = async () => {
-  const blockedDomains = await getRecord<string[]>(STORAGE_KEYS.BLOCKED_DOMAINS);
-  cachedBlockedDomains = blockedDomains ?? [];
-};
-
-export const getBlockedDomains = async () => {
-  if (cachedBlockedDomains) {
-    return cachedBlockedDomains;
-  }
-
-  await cacheBlockedDomains();
-  return cachedBlockedDomains;
-};
-
-export const isUrlInBlockList = async (url: string) => {
-  const blockedDomains = await getBlockedDomains();
-  return blockedDomains?.some((domain) => {
-    return matchSourceUrl(
-      {
-        key: SourceKey.HOST,
-        value: `/^(.+\.)?${domain}$/i`, // to match the domain and all its subdomains
-        operator: SourceOperator.MATCHES,
-      },
-      url
-    );
-  });
-};
-
-export const onBlockListChange = (callback: () => void) => {
-  onRecordChange<string[]>(
-    {
-      keyFilter: STORAGE_KEYS.BLOCKED_DOMAINS,
-      changeTypes: [ChangeType.MODIFIED],
-    },
-    () => {
-      cacheBlockedDomains().then(() => {
-        callback?.();
-      });
-    }
-  );
 };
